fix: proto generation using docker
Since commit 0664d8b1e3452b040f9e7cd448110fd3f90f842c there is
a bug in generation of the protobuf code: protoc-gen-go cannot
find launcher.proto file, as it is now not in internal/protobuf
directory. For this reason, the symlink was broken inside the
container.

This patch fixes the bug by changing mounting path of the container.
Also, generate.sh script was extended to be able to generate code with
docker (with option `USE_DOCKER=true`).